+++ AdjustAdobeExtensionApp/AdjustAdobeExtensionApp/SceneDelegate.h
@@ -1,0 +1,15 @@
+//
+//  SceneDelegate.h
+//  AdjustAdobeExtensionApp
+//
+//  Created by Adjust SDK Team on 21.11.23.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
+
+@property (strong, nonatomic) UIWindow * window;
+
+@end
+